requirements.txt
requests
jinja2
psycopg2-binary